+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `targetCellWidth` and `targetCellHeight` options in config plugin
+
 ## [0.13.2] - 2024-05-12
 
 ### Fixed